requirements.txt
opencv-python
mediapipe
numpy
requests
python-dotenv
openai
pillow
protobuf==3.20.3